grsec: Apply grsecurity-2.2.2-2.6.38.6-201105111839.patch
Changes in grsecurity-2.2.2-2.6.38.6-201105111839.patch:
---------------------------------------------------------------
update to latest PaX patch, increases i386 RANDKSTACK entropy

 arch/x86/ia32/ia32entry.S               |    2 +-
 arch/x86/include/asm/atomic64_32.h      |   31 +++++++++++++++++++++++++++++++
 arch/x86/include/asm/processor.h        |    6 +++---
 arch/x86/kernel/cpu/common.c            |    2 +-
 arch/x86/kernel/head_32.S               |    8 ++++----
 arch/x86/kernel/process.c               |    7 +++++--
 arch/x86/kernel/process_64.c            |    6 +++---
 arch/x86/kernel/pvclock.c               |    8 ++++----
 arch/x86/kernel/smpboot.c               |    2 +-
 arch/x86/mm/pgtable.c                   |   15 ++++-----------
 arch/x86/xen/smp.c                      |    2 +-
 drivers/infiniband/hw/ipath/ipath_rc.c  |    6 +++---
 drivers/infiniband/hw/ipath/ipath_ruc.c |    6 +++---
 drivers/net/tg3.c                       |    2 +-
 drivers/spi/spi.c                       |    2 +-
 fs/configfs/dir.c                       |   10 ++++++++--
 fs/gfs2/ops_inode.c                     |    2 +-
 fs/libfs.c                              |   13 +++++++------
 fs/ocfs2/suballoc.c                     |    2 +-
 fs/proc/meminfo.c                       |    2 --
 fs/select.c                             |    6 +++---
 fs/splice.c                             |    7 +++----
 include/linux/sched.h                   |   17 ++++++++++-------
 ipc/sem.c                               |    4 ++--
 mm/shmem.c                              |    3 +--
 net/ipv4/tcp.c                          |    6 ------
 net/ipv6/inet6_hashtables.c             |    2 +-
 net/rds/ib.h                            |    2 +-
 net/rds/ib_cm.c                         |    2 +-
 net/rds/ib_recv.c                       |    4 ++--
 net/rds/iw.h                            |    2 +-
 net/rds/iw_cm.c                         |    2 +-
 net/rds/iw_recv.c                       |    4 ++--
 33 files changed, 111 insertions(+), 84 deletions(-)
---------------------------------------------------------------

Signature-tree: 89ee035e0bb8167af7460cf89a072c058999133b

+++ mm/shmem.c
@@ -2343,8 +2343,7 @@ int shmem_fill_super(struct super_block *sb, void *data, int silent)
 	int err = -ENOMEM;
 
 	/* Round up to L1_CACHE_BYTES to resist false sharing */
-	sbinfo = kzalloc(max((int)sizeof(struct shmem_sb_info),
-				L1_CACHE_BYTES), GFP_KERNEL);
+	sbinfo = kzalloc(max(sizeof(struct shmem_sb_info), L1_CACHE_BYTES), GFP_KERNEL);
 	if (!sbinfo)
 		return -ENOMEM;
